Change 'Time' constraint to answer 1-2 (a)
Explanation: there’s no difference in the solution returned compared to
before changing the ‘Time’ constraint, because the first stage ‘reheat’
has a lower number of available hours than the second stage ‘roll’.

If for experiment’s sake we give the ‘reheat’ stage a higher number of
available hours than the ‘roll’ stage, this current wording of the
‘Time’ constraint causes AMPL to return the error:

  MINOS 5.51: infeasible problem.
  1 iterations

So, to find a more robust way to handle this I’m in search of some kind
of the-greater-of operand which I hope will emerge as we go deeper into
the book. An IF statement would do nicely.

Hmmm.. maybe that, or maybe I’m not seeing an obvious way to rejig the
constraint so that it’s resilient to a higher number of available hours
in earlier stages than later ones.