Add constraint 'Share'
This new constraint completes the replacement of param
‘commit’ with ‘min_share’ for use as a lower bound on the production
quantities for each product.

ALSO: I returned the ‘min_share' values for bands and plate back to 0.4
from 0.5, which I had changed initially to mess around with the later
part of the question, covered in the CHAPTER NOTES.md file in this
repository’s master branch.